pyproject.toml
[tool.poetry]
name = "tiktoken-grpc"
version = "0.1.0"
description = ""
authors = ["Yoriyasu Yano <430092+yorinasub17@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "tiktoken_grpc"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.3"
grpcio-tools = "^1.51.3"
tiktoken = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
